Added Assert (p 751 in The C++ Programming Language)
- Added unit tests for Assert
- Fixed test project setup

+++ rts-wars/rts-wars/graphics/cameras/SfmlCamera.cpp
@@ -12,7 +12,7 @@ const float movementPerTick = 0.1f;
 SfmlCamera::SfmlCamera(sf::RenderWindow& window, sf::Vector2u worldDimensions) :
 	window(window),
 	worldDimensions(worldDimensions),
-	view(sf::FloatRect(0, 0, window.getSize().x, window.getSize().y))
+	view(sf::FloatRect(0, 0, static_cast<float>(window.getSize().x), static_cast<float>(window.getSize().y)))
 {
 }
